rename documentstore -> document_store
