+++ haystack/document_store/base.py
