+++ test/test_eval.py
@@ -1,5 +1,5 @@
 import pytest
-from haystack.documentstore.base import BaseDocumentStore
+from haystack.document_store.base import BaseDocumentStore
 from haystack.retriever.sparse import ElasticsearchRetriever
 from haystack.finder import Finder
